+++ packages/api/src/external/fhir-to-cda-converter/connector-lambda.ts
@@ -2,7 +2,6 @@ import { Input } from "@metriport/core/domain/conversion/fhir-to-cda";
 import { getLambdaResultPayload, makeLambdaClient } from "@metriport/core/external/aws/lambda";
 import { getOrganizationOrFail } from "../../command/medical/organization/get-organization";
 import { Config } from "../../shared/config";
-import { toFHIR as toFhirOrganization } from "../fhir/organization";
 import { FhirToCdaConverter, FhirToCdaConverterRequest } from "./connector";
 
 const region = Config.getAWSRegion();
@@ -10,23 +9,15 @@ const lambdaClient = makeLambdaClient(region);
 const fhirToCdaConverterLambdaName = Config.getFhirToCdaConverterLambdaName();
 
 export class FhirToCdaConverterLambda implements FhirToCdaConverter {
-  async requestConvert({
-    cxId,
-    patientId,
-    docId,
-    bundle,
-  }: FhirToCdaConverterRequest): Promise<void> {
+  async requestConvert({ cxId, bundle, toSplit }: FhirToCdaConverterRequest): Promise<string[]> {
     if (!fhirToCdaConverterLambdaName) {
       throw new Error("FHIR to CDA Converter Lambda Name is undefined");
     }
     const organization = await getOrganizationOrFail({ cxId });
-    const fhirOrganization = toFhirOrganization(organization);
     const lambdaInput: Input = {
       cxId,
-      patientId,
-      docId,
       bundle,
-      organization: fhirOrganization,
+      toSplit,
       orgOid: organization.oid,
     };
 
@@ -38,13 +29,13 @@ export class FhirToCdaConverterLambda implements FhirToCdaConverter {
       })
       .promise();
 
-    // Intentionally not assigned. Used to check for errors in the lambda result.
-    getLambdaResultPayload({
+    // TODO: Check that this works on Staging
+    const res = getLambdaResultPayload({
       result,
       lambdaName: fhirToCdaConverterLambdaName,
       failGracefully: false,
     });
 
-    return;
+    return [res];
   }
 }
