+++ packages/api/src/external/fhir-to-cda-converter/connector.ts
@@ -2,11 +2,10 @@ import { Bundle } from "../../routes/medical/schemas/fhir";
 
 export type FhirToCdaConverterRequest = {
   cxId: string;
-  patientId: string;
-  docId: string;
+  toSplit: boolean;
   bundle: Bundle;
 };
 
 export interface FhirToCdaConverter {
-  requestConvert(req: FhirToCdaConverterRequest): Promise<void>;
+  requestConvert(req: FhirToCdaConverterRequest): Promise<string[]>;
 }
