+++ packages/lambdas/src/fhir-to-cda-converter.ts
@@ -1,6 +1,6 @@
 import { Input } from "@metriport/core/domain/conversion/fhir-to-cda";
+import { splitBundleByCompositions } from "@metriport/core/fhir-to-cda/composition-splitter";
 import { convertFhirBundleToCda } from "@metriport/core/fhir-to-cda/fhir-to-cda";
-import { uploadCdaDocuments } from "@metriport/core/fhir-to-cda/upload";
 import { out } from "@metriport/core/util/log";
 import * as Sentry from "@sentry/serverless";
 import { capture } from "./shared/capture";
@@ -13,20 +13,14 @@ const lambdaName = getEnvOrFail("AWS_LAMBDA_FUNCTION_NAME");
 const bucketName = getEnvOrFail("MEDICAL_DOCUMENTS_BUCKET_NAME");
 
 export const handler = Sentry.AWSLambda.wrapHandler(
-  async ({ cxId, patientId, docId, organization, bundle, orgOid }: Input): Promise<void> => {
-    const { log } = out(`cx ${cxId}, patient ${patientId}`);
+  async ({ cxId, bundle, orgOid, toSplit }: Input): Promise<string[]> => {
+    const { log } = out(`cx ${cxId}`);
     log(
       `Running with: ${bundle.entry?.length} resources, bundle type: ${bundle.type}, bucket: ${bucketName}}`
     );
     try {
-      const converted = convertFhirBundleToCda(bundle, orgOid);
-      await uploadCdaDocuments({
-        cxId,
-        patientId,
-        cdaBundles: converted,
-        organization,
-        docId,
-      });
+      const bundles = toSplit ? splitBundleByCompositions(bundle) : bundle;
+      return convertFhirBundleToCda(bundles, orgOid);
 
       //eslint-disable-next-line @typescript-eslint/no-explicit-any
     } catch (error: any) {
@@ -36,7 +30,6 @@ export const handler = Sentry.AWSLambda.wrapHandler(
         extra: {
           error,
           cxId,
-          patientId,
           context: lambdaName,
           orgOid,
         },
